+++ plugins/topology/src/components/Topology/TopologySideBar/TopologyResourcesTabPanel.tsx
@@ -16,6 +16,8 @@ import {
 } from '../../../models';
 import { JobData } from '../../../types/jobs';
 import PodStatus from '../../Pods/PodStatus';
+import PLRlist from './PLRlist';
+import { PodLogsDialog } from './PodLogs/PodLogsDialog';
 import IngressListSidebar from './Resources/IngressListSidebar';
 import RouteListSidebar from './Resources/RouteListSidebar';
 import TopologyResourcesTabPanelItem from './TopologyResourcesTabPaneltem';
@@ -30,6 +32,8 @@ const TopologyResourcesTabPanel = ({
   const data = node.getData();
   const nodeData = data?.data;
   const resource = data?.resource;
+  const pipelines = nodeData?.pipelinesData?.pipelines;
+  const pipelineRuns = nodeData?.pipelinesData?.pipelineRuns;
   const showIngressRoute = () => {
     const { ingressesData, routesData } = nodeData;
     const hasIngressData = ingressesData?.length > 0;
@@ -47,6 +51,7 @@ const TopologyResourcesTabPanel = ({
 
     return <IngressListSidebar ingressesData={ingressesData} />;
   };
+
   return (
     <div data-testid="resources-tab">
       <TopologyResourcesTabPanelItem
@@ -71,9 +76,15 @@ const TopologyResourcesTabPanel = ({
                   <Status status={pod.status?.phase ?? ''} />
                 </ResourceStatus>
               </span>
+              <span style={{ flex: '1' }}>
+                <PodLogsDialog podData={pod} />
+              </span>
             </li>
           ))}
       </TopologyResourcesTabPanelItem>
+      {pipelines?.length > 0 ? (
+        <PLRlist pipelines={pipelines} pipelineRuns={pipelineRuns} />
+      ) : null}
       {resource.kind === CronJobModel.kind ? (
         <TopologyResourcesTabPanelItem
           resourceLabel={JobModel.labelPlural}
@@ -129,12 +140,14 @@ const TopologyResourcesTabPanel = ({
                 {(service.spec?.ports ?? []).map(
                   ({ name, port, protocol, targetPort }: V1ServicePort) => (
                     <li key={name || `${port}-${protocol}`}>
-                      <span className="topology-text-muted">Service port:</span>{' '}
+                      <span className="bs-topology-text-muted">
+                        Service port:
+                      </span>{' '}
                       {name || `${port}-${protocol}`}
                       &nbsp;
                       <LongArrowAltRightIcon />
                       &nbsp;
-                      <span className="topology-text-muted">
+                      <span className="bs-topology-text-muted">
                         Pod port:
                       </span>{' '}
                       {targetPort}