+++ plugins/topology/src/data-transforms/data-transformer.ts
@@ -4,6 +4,7 @@ import { Model, NodeModel } from '@patternfly/react-topology';
 import { TYPE_APPLICATION_GROUP, TYPE_WORKLOAD } from '../const';
 import { CronJobModel } from '../models';
 import { K8sResponseData, K8sWorkloadResource } from '../types/types';
+import { getPipelinesAndPipelineRunsForResource } from '../utils/pipeline-utils';
 import { getPodsDataForResource } from '../utils/pod-resource-utils';
 import {
   createOverviewItemForType,
@@ -64,6 +65,10 @@ export const getBaseTopologyDataModel = (resources: K8sResponseData): Model => {
                     jobsData: getJobsDataForResource(resources, resource),
                   }
                 : {}),
+              pipelinesData: getPipelinesAndPipelineRunsForResource(
+                resource,
+                resources,
+              ),
             },
           );
           typedDataModel.nodes?.push(