+++ plugins/topology/src/types/types.ts
@@ -58,4 +58,9 @@ export type K8sResourcesContextData = {
   selectedClusterErrors?: ClusterErrors;
   clusters: string[];
   setSelectedCluster: React.Dispatch<React.SetStateAction<number>>;
+  selectedCluster?: number;
+};
+
+export type TektonResponseData = {
+  [key: string]: { data: any[] };
 };